# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projetlvre",
        "description": "une sorte de biblio en ligne",
        "version": "1.0"
    },
    "paths": {
...
        "/api/addcart/{id}": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add a book to the user's shopping cart",
                "description": "Add a book to the cart",
                "operationId": "96a48541f74b8f5e456f9a3ad449fd65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 12
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Book successfully added to cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Book added to cart successfully"
                                        },
                                        "cart": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Retrieve all books currently in the user cart",
                "description": "View user cart",
                "operationId": "b59a70997dfdd532dad7d6eef7e19437",
                "responses": {
                    "200": {
                        "description": "Cart retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "booksCart": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "Amount": {
                                            "type": "number",
                                            "example": 54.9
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Books in cart retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/deletelivrecart/{id}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Remove a specific book from the user's cart",
                "description": "Remove a book from the cart",
                "operationId": "d4b4944ba510577f8434d04843df4307",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cart item ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book removed from cart successfully"
                    },
                    "404": {
                        "description": "item not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}